- It puts the Widget in the basket.
