+++ lib/questionnaires/view/item/item.dart
@@ -4,5 +4,6 @@ export 'src/group_item.dart';
 export 'src/item_media_image.dart';
 export 'src/question_response_item_filler.dart';
 export 'src/questionnaire_item_filler.dart';
+export 'src/questionnaire_item_filler_title.dart';
 export 'src/response_item_filler.dart';
 export 'src/total_score_item.dart';
