+++ lib/questionnaires/view/view.dart
@@ -13,6 +13,8 @@ export 'src/questionnaire_information_tile.dart';
 export 'src/questionnaire_loading_indicator.dart';
 export 'src/questionnaire_page_scaffold.dart';
 export 'src/questionnaire_scroller.dart';
+export 'src/questionnaire_scroller_page.dart';
 export 'src/questionnaire_stepper.dart';
+export 'src/questionnaire_stepper_page.dart';
 export 'src/questionnaire_theme.dart';
 export 'src/xhtml.dart';
